+++ samila/params.py
@@ -41,6 +41,7 @@ PLOT_DATA_ERROR = "Plotting process can't be Done because data{0} is empty. Use 
 SAVE_NO_DATA_ERROR = "Data file can't be saved. At least one of the data1 or data2 is None."
 MATPLOTLIB_VERSION_WARNING = "Source matplotlib version({0}) is different from yours, plots may be different."
 CALCULATION_EXCEPTION_WARNING = "The given functions are undefined at some points. Your plot may not be complete."
+BOTH_COLOR_COMPLEMENT_WARNING = "Both color and bgcolor are 'complement'. Both are set to default."
 
 
 class Projection(Enum):